Create JumpGame2.java
Sample Input 1:
5
2 3 1 1 4
Sample Output 1:
2
Explanation of sample input 1 :

Consider the above figure:
We are initially at index 0, ARR[0] is 2 which represents we can jump a maximum of 2 steps. 

We jump 1 index from 0 to 1. At index 1, 'ARR[1]' is 3 which represents we can jump a maximum of 3 steps so we jump 3 indices from 1 to 4 to reach the last index. Hence we return 2.

It can be proved that the end can't be reached in less than 2.
Sample Input 2:
5
3 2 1 0 1
Sample Output 2:
-1